add .hpp in obstacle detection
